backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.34.0
anthropic>=0.52.0
tavily-python>=0.5.0
asyncpg>=0.30.0
slowapi>=0.1.9
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
